Fix zooming annotations in interactive mode
In seekwatcher, if you decide to select a rectangle for zooming, for
example, you will get an annotation box for the point where you release
the mouse button.  For me, this wasn't the expected behaviour.  To
remedy that, I just checked to see that the x and y position of the
mouse is the same at button release time as it was on button press.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>