Rewrite plotting routines.  Faster, much less ram.
The old code would create an array of all the data from the trace,
and then filter that array over and over again to produce the
different graphs.

The new code creates only the array needed to do the data plot,
and it tries to be as small as possible.

The data for the other graphs is created as we read the blkparse output,
so we don't have to iterate over all the blktrace data more than
once.