cmd/seekwatcher: switch from `mencoder` to direct `ffmpeg` call
`mencoder` is a part of `mplayer` package which does not get updated
or even packages in distributions nowadays.

The change switched to use of `ffmpeg` to encode video. While at it
switch from `mpeg2` to `x264`. At least for `ffmpeg-6` that shrinks
video size from 6MB down to 920KB of a sample trace I have locally.

Dropped `--movie-vbitrate` option as it's not used anymore.